+++ resolvers/webpack/index.js
@@ -11,6 +11,7 @@ const isCore = require('is-core-module');
 const resolve = require('resolve');
 const semver = require('semver');
 const has = require('has');
+const isRegex = require('is-regex');
 
 const log = require('debug')('eslint-plugin-import:resolver:webpack');
 
@@ -125,21 +126,28 @@ exports.resolve = function (source, file, settings) {
     }
   }
 
+  if (typeof webpackConfig.then === 'function') {
+    webpackConfig = {};
+
+    console.warn('Webpack config returns a `Promise`; that signature is not supported at the moment. Using empty object instead.');
+  }
+
   if (webpackConfig == null) {
     webpackConfig = {};
 
-    console.warn('No webpack configuration with a "resolve" field found. Using empty object instead');
+    console.warn('No webpack configuration with a "resolve" field found. Using empty object instead.');
   }
 
   log('Using config: ', webpackConfig);
 
+  const resolveSync = getResolveSync(configPath, webpackConfig, cwd);
+
   // externals
-  if (findExternal(source, webpackConfig.externals, path.dirname(file))) {
+  if (findExternal(source, webpackConfig.externals, path.dirname(file), resolveSync)) {
     return { found: true, path: null };
   }
 
   // otherwise, resolve "normally"
-  const resolveSync = getResolveSync(configPath, webpackConfig, cwd);
 
   try {
     return { found: true, path: resolveSync(path.dirname(file), source) };
@@ -243,19 +251,15 @@ function createWebpack1ResolveSync(webpackRequire, resolveConfig, plugins) {
   const SyncNodeJsInputFileSystem = webpackRequire('enhanced-resolve/lib/SyncNodeJsInputFileSystem');
 
   const ModuleAliasPlugin = webpackRequire('enhanced-resolve/lib/ModuleAliasPlugin');
-  const ModulesInDirectoriesPlugin =
-    webpackRequire('enhanced-resolve/lib/ModulesInDirectoriesPlugin');
+  const ModulesInDirectoriesPlugin = webpackRequire('enhanced-resolve/lib/ModulesInDirectoriesPlugin');
   const ModulesInRootPlugin = webpackRequire('enhanced-resolve/lib/ModulesInRootPlugin');
   const ModuleAsFilePlugin = webpackRequire('enhanced-resolve/lib/ModuleAsFilePlugin');
   const ModuleAsDirectoryPlugin = webpackRequire('enhanced-resolve/lib/ModuleAsDirectoryPlugin');
-  const DirectoryDescriptionFilePlugin =
-    webpackRequire('enhanced-resolve/lib/DirectoryDescriptionFilePlugin');
-  const DirectoryDefaultFilePlugin =
-    webpackRequire('enhanced-resolve/lib/DirectoryDefaultFilePlugin');
+  const DirectoryDescriptionFilePlugin = webpackRequire('enhanced-resolve/lib/DirectoryDescriptionFilePlugin');
+  const DirectoryDefaultFilePlugin = webpackRequire('enhanced-resolve/lib/DirectoryDefaultFilePlugin');
   const FileAppendPlugin = webpackRequire('enhanced-resolve/lib/FileAppendPlugin');
   const ResultSymlinkPlugin = webpackRequire('enhanced-resolve/lib/ResultSymlinkPlugin');
-  const DirectoryDescriptionFileFieldAliasPlugin =
-    webpackRequire('enhanced-resolve/lib/DirectoryDescriptionFileFieldAliasPlugin');
+  const DirectoryDescriptionFileFieldAliasPlugin = webpackRequire('enhanced-resolve/lib/DirectoryDescriptionFileFieldAliasPlugin');
 
   const resolver = new Resolver(new SyncNodeJsInputFileSystem());
 
@@ -320,30 +324,62 @@ function makeRootPlugin(ModulesInRootPlugin, name, root) {
 }
 /* eslint-enable */
 
-function findExternal(source, externals, context) {
+function findExternal(source, externals, context, resolveSync) {
   if (!externals) return false;
 
   // string match
   if (typeof externals === 'string') return (source === externals);
 
   // array: recurse
-  if (externals instanceof Array) {
-    return externals.some(function (e) { return findExternal(source, e, context); });
+  if (Array.isArray(externals)) {
+    return externals.some(function (e) { return findExternal(source, e, context, resolveSync); });
   }
 
-  if (externals instanceof RegExp) {
+  if (isRegex(externals)) {
     return externals.test(source);
   }
 
   if (typeof externals === 'function') {
     let functionExternalFound = false;
-    externals.call(null, context, source, function(err, value) {
+    const callback = function (err, value) {
       if (err) {
         functionExternalFound = false;
       } else {
-        functionExternalFound = findExternal(source, value, context);
+        functionExternalFound = findExternal(source, value, context, resolveSync);
       }
-    });
+    };
+    // - for prior webpack 5, 'externals function' uses 3 arguments
+    // - for webpack 5, the count of arguments is less than 3
+    if (externals.length === 3) {
+      externals.call(null, context, source, callback);
+    } else {
+      const ctx = {
+        context,
+        request: source,
+        contextInfo: {
+          issuer: '',
+          issuerLayer: null,
+          compiler: '',
+        },
+        getResolve: () => (resolveContext, requestToResolve, cb) => {
+          if (cb) {
+            try {
+              cb(null, resolveSync(resolveContext, requestToResolve));
+            } catch (e) {
+              cb(e);
+            }
+          } else {
+            log('getResolve without callback not supported');
+            return Promise.reject(new Error('Not supported'));
+          }
+        },
+      };
+      const result = externals.call(null, ctx, callback);
+      // todo handling Promise object (using synchronous-promise package?)
+      if (result && typeof result.then === 'function') {
+        log('Asynchronous functions for externals not supported');
+      }
+    }
     return functionExternalFound;
   }