+++ src/rules/extensions.js
@@ -2,6 +2,7 @@ import path from 'path';
 
 import resolve from 'eslint-module-utils/resolve';
 import { isBuiltIn, isExternalModule, isScoped, isScopedModule } from '../core/importType';
+import moduleVisitor from 'eslint-module-utils/moduleVisitor';
 import docsUrl from '../docsUrl';
 
 const enumValues = { enum: [ 'always', 'ignorePackages', 'never' ] };
@@ -129,17 +130,15 @@ module.exports = {
     function isExternalRootModule(file) {
       const slashCount = file.split('/').length - 1;
 
+      if (slashCount === 0)  return true;
       if (isScopedModule(file) && slashCount <= 1) return true;
-      if (isExternalModule(file, context, resolve(file, context)) && !slashCount) return true;
       return false;
     }
 
-    function checkFileExtension(node) {
-      const { source } = node;
-
-      // bail if the declaration doesn't have a source, e.g. "export { foo };"
-      if (!source) return;
-
+    function checkFileExtension(source) {
+      // bail if the declaration doesn't have a source, e.g. "export { foo };", or if it's only partially typed like in an editor
+      if (!source || !source.value) return;
+      
       const importPathWithQueryString = source.value;
 
       // don't enforce anything on builtins
@@ -158,8 +157,12 @@ module.exports = {
       const extension = path.extname(resolvedPath || importPath).substring(1);
 
       // determine if this is a module
-      const isPackage = isExternalModule(importPath, context.settings)
-        || isScoped(importPath);
+      const isPackage = isExternalModule(
+        importPath,
+        context.settings,
+        resolve(importPath, context),
+        context
+      ) || isScoped(importPath);
 
       if (!extension || !importPath.endsWith(`.${extension}`)) {
         const extensionRequired = isUseOfExtensionRequired(extension, isPackage);
@@ -181,9 +184,6 @@ module.exports = {
       }
     }
 
-    return {
-      ImportDeclaration: checkFileExtension,
-      ExportNamedDeclaration: checkFileExtension,
-    };
+    return moduleVisitor(checkFileExtension, { commonjs: true });
   },
 };