+++ src/rules/no-internal-modules.js
@@ -2,7 +2,7 @@ import minimatch from 'minimatch';
 
 import resolve from 'eslint-module-utils/resolve';
 import importType from '../core/importType';
-import isStaticRequire from '../core/staticRequire';
+import moduleVisitor from 'eslint-module-utils/moduleVisitor';
 import docsUrl from '../docsUrl';
 
 module.exports = {
@@ -14,16 +14,32 @@ module.exports = {
 
     schema: [
       {
-        type: 'object',
-        properties: {
-          allow: {
-            type: 'array',
-            items: {
-              type: 'string',
+        oneOf: [
+          {
+            type: 'object',
+            properties: {
+              allow: {
+                type: 'array',
+                items: {
+                  type: 'string',
+                },
+              },
             },
+            additionalProperties: false,
           },
-        },
-        additionalProperties: false,
+          {
+            type: 'object',
+            properties: {
+              forbid: {
+                type: 'array',
+                items: {
+                  type: 'string',
+                },
+              },
+            },
+            additionalProperties: false,
+          },
+        ],
       },
     ],
   },
@@ -31,11 +47,7 @@ module.exports = {
   create: function noReachingInside(context) {
     const options = context.options[0] || {};
     const allowRegexps = (options.allow || []).map(p => minimatch.makeRe(p));
-
-    // test if reaching to this destination is allowed
-    function reachingAllowed(importPath) {
-      return allowRegexps.some(re => re.test(importPath));
-    }
+    const forbidRegexps = (options.forbid || []).map(p => minimatch.makeRe(p));
 
     // minimatch patterns are expected to use / path separators, like import
     // statements, so normalize paths to use the same
@@ -43,9 +55,8 @@ module.exports = {
       return somePath.split('\\').join('/');
     }
 
-    // find a directory that is being reached into, but which shouldn't be
-    function isReachViolation(importPath) {
-      const steps = normalizeSep(importPath)
+    function toSteps(somePath) {
+      return  normalizeSep(somePath)
         .split('/')
         .reduce((acc, step) => {
           if (!step || step === '.') {
@@ -56,6 +67,20 @@ module.exports = {
             return acc.concat(step);
           }
         }, []);
+    }
+
+    // test if reaching to this destination is allowed
+    function reachingAllowed(importPath) {
+      return allowRegexps.some(re => re.test(importPath));
+    }
+
+    // test if reaching to this destination is forbidden
+    function reachingForbidden(importPath) {
+      return forbidRegexps.some(re => re.test(importPath));
+    }
+
+    function isAllowViolation(importPath) {
+      const steps = toSteps(importPath);
 
       const nonScopeSteps = steps.filter(step => step.indexOf('@') !== 0);
       if (nonScopeSteps.length <= 1) return false;
@@ -75,6 +100,27 @@ module.exports = {
       return true;
     }
 
+    function isForbidViolation(importPath) {
+      const steps = toSteps(importPath);
+
+      // before trying to resolve, see if the raw import (with relative
+      // segments resolved) matches a forbidden pattern
+      const justSteps = steps.join('/');
+
+      if (reachingForbidden(justSteps) || reachingForbidden(`/${justSteps}`)) return true;
+
+      // if the import statement doesn't match directly, try to match the
+      // resolved path if the import is resolvable
+      const resolved = resolve(importPath, context);
+      if (resolved && reachingForbidden(normalizeSep(resolved))) return true;
+
+      // this import was not forbidden by the forbidden paths so it is not a violation
+      return false;
+    }
+
+    // find a directory that is being reached into, but which shouldn't be
+    const isReachViolation = options.forbid ? isForbidViolation : isAllowViolation;
+
     function checkImportForReaching(importPath, node) {
       const potentialViolationTypes = ['parent', 'index', 'sibling', 'external', 'internal'];
       if (potentialViolationTypes.indexOf(importType(importPath, context)) !== -1 &&
@@ -87,24 +133,8 @@ module.exports = {
       }
     }
 
-    return {
-      ImportDeclaration(node) {
-        checkImportForReaching(node.source.value, node.source);
-      },
-      ExportAllDeclaration(node) {
-        checkImportForReaching(node.source.value, node.source);
-      },
-      ExportNamedDeclaration(node) {
-        if (node.source) {
-          checkImportForReaching(node.source.value, node.source);
-        }
-      },
-      CallExpression(node) {
-        if (isStaticRequire(node)) {
-          const [ firstArgument ] = node.arguments;
-          checkImportForReaching(firstArgument.value, firstArgument);
-        }
-      },
-    };
+    return moduleVisitor((source) => {
+      checkImportForReaching(source.value, source);
+    }, { commonjs: true });
   },
 };